Conflicts:
	code/src/client/actors/StandardCellClient.cpp
	code/src/client/actors/StemCellClient.cpp

+++ code/src/client/actors/GameObjectClient.cpp
@@ -15,6 +15,6 @@ void GameObjectClient::draw() const
 
 void GameObjectClient::drawAtTransformation() const
 {
-	gl::color(ColorA(1, 0, 0, opacity));
-	drawSphere(Vec3f(0, 0, 0), radius, radius);
+	gl::color(ColorA(1.f, 0.f, 0.f, opacity));
+	drawSphere(Vec3f(0.f, 0.f, 0.f), radius, (int) radius);
 }
